refactor, fuzz: Split ChaCha20SplitFuzz to fuzz crypto utils
